02_clientes/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Django==3.2.10
django-filter==22.1
djangorestframework==3.13.1
Faker==13.15.1
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
validate-docbr==1.10.0
